-fixed bugs

+++ bot.py
@@ -240,22 +240,35 @@ def time_pick(information):
 
 
 def buy(information):
-    bot.delete_message(information.message.chat.id, information.message.message_id)
-    bot.send_invoice(
-        information.message.chat.id,
-        title='Продолжение курса',
-        description='Чтобы продолжить обучение, заплатите за следующие 12 вебинаров',
-        provider_token=PAYMENTS_PROVIDER_TOKEN,
-        currency='RUB',
-        photo_url='https://www.instituteiba.by/upload/medialibrary/5f5/5f5aa75c5497429160440528683d411c.jpg',
-        photo_height=560,
-        photo_width=1024,
-        photo_size=512,
-        is_flexible=False,  # True, если конечная цена зависит от способа доставки
-        prices=[PRICE],
-        start_parameter='paying_webinars',
-        invoice_payload='webinars_payed'
-    )
+    cursor.execute('SELECT * FROM Users')
+
+    data_arr = cursor.fetchall()
+
+    if len(data_arr) > 0:
+        for data in data_arr:
+            if data[0] == information.from_user.id:
+
+                if data[2] == 13 | data[2] == 0:
+                    bot.delete_message(information.message.chat.id, information.message.message_id)
+                    bot.send_invoice(
+                        information.message.chat.id,
+                        title='Продолжение курса',
+                        description='Чтобы продолжить обучение, заплатите за следующие 12 вебинаров',
+                        provider_token=PAYMENTS_PROVIDER_TOKEN,
+                        currency='RUB',
+                        photo_url='https://www.instituteiba.by/upload/medialibrary/5f5/5f5aa75c5497429160440528683d411c.jpg',
+                        photo_height=560,
+                        photo_width=1024,
+                        photo_size=512,
+                        is_flexible=False,  # True, если конечная цена зависит от способа доставки
+                        prices=[PRICE],
+                        start_parameter='paying_webinars',
+                        invoice_payload='webinars_payed'
+                    )
+                    break
+        else:
+            bot.send_message(information.message.chat.id, 'Вы уже оплатили занятия и сможете снова за них заплатить,'
+                                                          'когда пройдёте их все до конца')
 
 
 @bot.pre_checkout_query_handler(func=lambda query: True)
